+++ MulliganMadness/Cards/CardRegistration.cs
@@ -6,8 +6,6 @@ namespace MulliganMadness.Cards
     {
         internal static void RegisterAll()
         {
-            CustomCard.BuildCard<NestEgg>(info => NestEgg.Card = info);
-            CustomCard.BuildCard<SilverEgg>(info => SilverEgg.Card = info);
             CustomCard.BuildCard<ReturnToSender>(info => ReturnToSender.Card = info);
         }
     }